# Conflicts:
#	projectbob/src/main/java/com/projectbob/controller/BobController.java
#	projectbob/src/main/resources/application.properties
#	projectbob/src/main/resources/templates/layouts/main_layout.html

+++ projectbob/src/main/java/com/projectbob/domain/Menu.java
@@ -1,5 +1,6 @@
 package com.projectbob.domain;
 
+import java.math.BigDecimal;
 import java.sql.*;
 import java.util.*;
 
@@ -24,13 +25,17 @@ public class Menu {
 
 	private List<MenuOption> options;
 	
-	private Double calories;
-    private Double carbs;
-    private Double protein;
-    private Double fat;
-    private Double sfa;
-    private Double sugar;
-    private Double sodium;
-    private Integer servingSize;
+
+	 private BigDecimal calories; // 열량
+	 private BigDecimal carbs;    // 탄수화물
+	 private BigDecimal protein;  // 단백질
+	 private BigDecimal fat;      // 지방
+	 private BigDecimal sfa;      // 포화지방
+	 private BigDecimal sugar;    // 당
+	 private BigDecimal sodium;   // 나트륨
+	 private BigDecimal servingSize; // 제공량
+
+
+
 
 }
